create search component, moved code from navbar to search
