+++ react-app/src/components/NavBar/NavBar.js
@@ -19,13 +19,7 @@ const NavBar = () => {
   // const [searchText, setSearchText] = useState('');
 
 
-  //Search Props
-  const [searchText, setSearchText] = useState('')
 
-  const searchMeFunc = (e) => {
-    e.preventDefault();
-    setSearchText(e.target.value)
-  }
 
 
 
@@ -57,18 +51,7 @@ const NavBar = () => {
           </Link>
         </div>
 
-        <div className='search-div' >
-          <div>
-        <input
-          placeholder="Search"
-          type="text"
-          onChange={searchMeFunc}
-          value={searchText}
-        />
-        <button type="submit" className='search-button' onClick={(e) => searchMeFunc()}
-><i class="fa-solid fa-magnifying-glass"></i></button>
-        </div>
-        </div>
+
 
         <div className="user-icon-wrapper">
         <NavLink
